requirements.txt
prometheus-client>=0.15.0
requests>=2.32.3
python-dotenv>=1.0